[Refactor] 클로저 내부 weak self 사용, products 일반 프로퍼티로 변경

+++ BuyOrNot/BuyOrNot/Controller/RankManager.swift
@@ -12,7 +12,8 @@ final class RankManager {
     static let shared = RankManager()
 
     func refreshRank() {
-        DispatchQueue.global().async {
+        DispatchQueue.global().async { [weak self] in
+            guard let self = self else { return }
             self.deleteAll()
             ProductCategory.allCases.forEach {
                 DanawaCrawler.shared.fetchRankData(category: $0).forEach { product in
